requirements.txt
inferactively-pymdp
matplotlib